Clean unused views and helpers
