--- app/views/proposals/_proposal.html.erb
@@ -1,61 +1,0 @@
-<div id="<%= dom_id(proposal) %>" class="proposal clear <%= proposal_class_names proposal %>" data-timestamp="<%= proposal.created_at.to_i %>">
-    <div class="row">
-      <div class="small-12 medium-9 column">
-        <%= cache [proposal, proposal.author, I18n.locale] do %>
-          <div class="proposal-content">
-            <%= proposal_badge(proposal) %>
-            <span class="label-proposal"><%= t("proposals.proposal.proposal") %></span>
-            <h3><%= link_to truncate(proposal.title, length: 100), namespaced_proposal_path(proposal) %></h3>
-            <p class="proposal-info">
-              <%= l proposal.created_at.to_date %>
-
-              <% unless proposal.official? || proposal.from_meeting? %>
-                <% if proposal.author.hidden? || proposal.author.erased? %>
-                  <span class="bullet">&nbsp;&bull;&nbsp;</span>
-                  <span class="author">
-                    <%= t("proposals.show.author_deleted") %>
-                  </span>
-                <% else %>
-                  <span class="bullet">&nbsp;&bull;&nbsp;</span>
-                  <span class="author">
-                    <%= link_to user_path(proposal.author) do %>
-                      <%= proposal.author.name %>
-                    <% end %>
-                  </span>
-                  <% if proposal.author.official? %>
-                    <span class="bullet">&nbsp;&bull;&nbsp;</span>
-                    <span class="label round level-<%= proposal.author.official_level %>">
-                      <%= official_position(proposal.author) %>
-                    </span>
-                  <% end %>
-                <% end %>
-
-                <% if proposal.author.verified_organization? %>
-                  <span class="bullet">&nbsp;&bull;&nbsp;</span>
-                  <span class="label round is-association">
-                    <%= t("shared.collective") %>
-                  </span>
-                <% end %>
-              <% end %>
-            </p>
-            <div class="proposal-description">
-              <p>
-                <%= truncate strip_tags(proposal.summary), length: 300 %>
-              </p>
-              <div class="truncate"></div>
-            </div>
-            <div class="bottom-bar">
-              <%= render "proposals/meta", proposal: proposal %>
-              <% if feature?(:proposal_tags) %>
-                <%= render "shared/tags", taggable: proposal, limit: 5 %>
-              <% end %>
-            </div>
-          </div>
-        <% end %>
-      </div>
-
-      <aside id="<%= dom_id(proposal) %>_votes" class="small-12 medium-3 column">
-        <%= vote_component_for proposal %>
-      </aside>
-    </div>
-</div>
