--- app/views/proposals/_proposals.html.erb
@@ -1,8 +1,0 @@
-<%= render 'shared/order_links', i18n_namespace: "proposals.index" %>
-
-<div class="show-for-small-only">
-  <%= link_to t("proposals.index.start_proposal"), new_proposal_path, class: 'button radius expand new-proposal' %>
-</div>
-
-<%= render partial: 'proposals/proposal', collection: @proposals %>
-<%= paginate @proposals, params: { random_seed: @random_seed } %>
