--- app/views/proposals/index.html.erb.bak
@@ -1,48 +1,0 @@
-<% provide :title do %><%= t('proposals.index.title') %><% end %>
-
-<section role="main">
-  <div class="page-title">
-    <div class="wrap row proposals-header">
-      <div class="small-12 columns">
-        <h2><%= t '.title' %></h2>
-      </div>
-    </div>
-    <div class="wrap row proposals-header">
-      <div class="small-12 medium-7 large-9 columns">
-        <h3><%= t '.description' %></h3>
-      </div>
-      <div class="small-12 medium-5 large-3 columns">
-        <div class="download-buttons">
-          <%= link_to page_path('download'), class: 'download-pam' do %>
-            <span><%= t('.download') %></span>
-          <% end %>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="wrap row">
-    <%= proposal_filter_tabs filter: @filter %>
-  </div>
-
-  <div class="wrap row">
-    <div class="small-12 medium-3 column">
-      <aside class="sidebar" role="complementary">
-        <%= link_to t("proposals.index.start_proposal"), new_proposal_path, class: 'new-proposal button radius expand' %>
-        <div class="sidebar-section proposal-filter-menu collapsed">
-          <h2 class="title"><%= t 'proposals.index.filter.title' %> <span class="toggle-menu"></span></h2>
-          <div class="sidebar-section-content">
-            <%= proposal_filters filter: @filter, tag_cloud: @tag_cloud %>
-          </div>
-        </div>
-      </aside>
-    </div>
-
-    <div class="small-12 medium-9 column">
-      <%= react_component('Loading') %>
-      <div id="proposals" class="proposals-list">
-        <%= render "proposals" %>
-      </div>
-    </div>
-  </div>
-</section>
