--- app/views/proposals/index.js.erb
@@ -1,4 +1,0 @@
-$('#proposals').html('<%= j render("proposals") %>');
-App.LocationChanger.initialize();
-$(document).trigger('tags:reload', '<%= @tag_cloud.to_json.html_safe %>');
-ReactRailsUJS.mountComponents();
